Add Bambino support to makefile
Adding Bambino200E and Bambino210E support to the makefile. By default
the makefile will still build the binary suitable for running on the
Smoothie2 hardware with its LPC4337 processor. Users with a Bambino
board can modify the following line in the makefile to match their
board:
    BOARD ?= Smoothie2
Instead of modifying the makefile, they can also specify the correct
board on the command line every time they build the binary. For
example:
  make clean all BOARD=Bambino210E

Also updates SerialConsole::SerialConsole() constructor to take the
tx/rx pin designations in the same order as the mbed SDK and as the
existing Smoothie code (ie. Kernel::Kernel()) expected as well.