Re-enable compiler optimizations
Commit 90ac0c1b2625e6e79842962fce332666a4ef4b8b disabled
optimizations. This made it easier to debug the code but it can make
code like the step ticker run so slowly that it eats all CPU cycles.
This was observed on the Bambino210E where the code is running out of
SPIFI FLASH.

I also force MRI_SEMIHOST_STDIO to 0 since we have never wanted this
feature enabled in Smoothie firmware.